requirement.txt
MySQLdb
python-docx
xlrd
xlwt
reportlab
